requirements.txt
selenium
webdriver-manager
requests
pandas
tqdm
python-dotenv
